+++ PlanIt-application/src/main/kotlin/com/nodoubt/planitbackend/domain/changeMaster/api/QueryReplacementClassDetailsPort.kt
@@ -1,0 +1,7 @@
+package com.nodoubt.planitbackend.domain.changeMaster.api
+
+import com.nodoubt.planitbackend.domain.changeMaster.api.dto.response.QueryReplacementClassDetailsResponse
+
+interface QueryReplacementClassDetailsPort {
+    fun execute(replacementClassId: Long): QueryReplacementClassDetailsResponse
+}
