+++ PlanIt-application/src/main/kotlin/com/nodoubt/planitbackend/domain/changeMaster/api/QueryReplacementClassListPort.kt
@@ -1,7 +1,7 @@
-package com.nodoubt.planitbackend.domain.replacementClass.api
+package com.nodoubt.planitbackend.domain.changeMaster.api
 
 import com.nodoubt.planitbackend.domain.changeMaster.domain.Status
-import com.nodoubt.planitbackend.domain.replacementClass.api.dto.response.QueryReplacementClassListResponse
+import com.nodoubt.planitbackend.domain.changeMaster.api.dto.response.QueryReplacementClassListResponse
 
 interface QueryReplacementClassListPort {
     fun execute(status: Status?): QueryReplacementClassListResponse
