+++ PlanIt-application/src/main/kotlin/com/nodoubt/planitbackend/domain/changeMaster/api/dto/response/QueryReplacementClassListResponse.kt
@@ -1,0 +1,7 @@
+package com.nodoubt.planitbackend.domain.changeMaster.api.dto.response
+
+import com.nodoubt.planitbackend.domain.changeMaster.spi.replacementClassSpi.vo.ReplacementClassListVO
+
+data class QueryReplacementClassListResponse(
+    val replacementClassList: List<ReplacementClassListVO>
+)
