+++ PlanIt-application/src/main/kotlin/com/nodoubt/planitbackend/domain/changeMaster/spi/reinforcementClassSpi/CommandReinforcementPort.kt
@@ -1,0 +1,8 @@
+package com.nodoubt.planitbackend.domain.changeMaster.spi.reinforcementClassSpi
+
+import com.nodoubt.planitbackend.domain.changeMaster.domain.ReinforcementClass
+
+interface CommandReinforcementPort {
+
+    fun saveReinforcementClassList(reinforcementClass: List<ReinforcementClass>)
+}
