+++ PlanIt-application/src/main/kotlin/com/nodoubt/planitbackend/domain/changeMaster/spi/reinforcementClassSpi/ReinforcementClassPort.kt
@@ -1,0 +1,6 @@
+package com.nodoubt.planitbackend.domain.changeMaster.spi.reinforcementClassSpi
+
+import com.nodoubt.planitbackend.domain.changeMaster.spi.reinforcementClassSpi.CommandReinforcementPort
+
+interface ReinforcementClassPort : CommandReinforcementPort {
+}
