+++ PlanIt-application/src/main/kotlin/com/nodoubt/planitbackend/domain/changeMaster/spi/replacementClassSpi/CommandReplacementPort.kt
@@ -1,0 +1,7 @@
+package com.nodoubt.planitbackend.domain.changeMaster.spi.replacementClassSpi
+
+import com.nodoubt.planitbackend.domain.changeMaster.domain.ReplacementClass
+
+interface CommandReplacementPort {
+    fun saveReplacementClassList(replacementClass: List<ReplacementClass>)
+}
