+++ PlanIt-application/src/main/kotlin/com/nodoubt/planitbackend/domain/changeMaster/spi/replacementClassSpi/QueryReplacementClassPort.kt
@@ -1,8 +1,8 @@
-package com.nodoubt.planitbackend.domain.replacementClass.spi
+package com.nodoubt.planitbackend.domain.changeMaster.spi.replacementClassSpi
 
 import com.nodoubt.planitbackend.domain.changeMaster.domain.Status
-import com.nodoubt.planitbackend.domain.replacementClass.spi.vo.ReplacementClassDetailsVO
-import com.nodoubt.planitbackend.domain.replacementClass.spi.vo.ReplacementClassListVO
+import com.nodoubt.planitbackend.domain.changeMaster.spi.replacementClassSpi.vo.ReplacementClassDetailsVO
+import com.nodoubt.planitbackend.domain.changeMaster.spi.replacementClassSpi.vo.ReplacementClassListVO
 
 interface QueryReplacementClassPort {
     fun queryReplacementClassListByUserId(userId: Long, status: Status?): List<ReplacementClassListVO>
