--- PlanIt-application/src/main/kotlin/com/nodoubt/planitbackend/domain/reinforcementClass/spi/CommandReinforcementPort.kt
@@ -1,8 +1,0 @@
-package com.nodoubt.planitbackend.domain.reinforcementClass.spi
-
-import com.nodoubt.planitbackend.domain.reinforcementClass.domain.ReinforcementClass
-
-interface CommandReinforcementPort {
-
-    fun saveReinforcementClassList(reinforcementClass: List<ReinforcementClass>)
-}
