--- PlanIt-application/src/main/kotlin/com/nodoubt/planitbackend/domain/reinforcementClass/spi/ReinforcementClassPort.kt
@@ -1,4 +1,0 @@
-package com.nodoubt.planitbackend.domain.reinforcementClass.spi
-
-interface ReinforcementClassPort : CommandReinforcementPort {
-}
