--- PlanIt-application/src/main/kotlin/com/nodoubt/planitbackend/domain/replacementClass/api/QueryReplacementClassDetailsPort.kt
@@ -1,7 +1,0 @@
-package com.nodoubt.planitbackend.domain.replacementClass.api
-
-import com.nodoubt.planitbackend.domain.replacementClass.api.dto.response.QueryReplacementClassDetailsResponse
-
-interface QueryReplacementClassDetailsPort {
-    fun execute(replacementClassId: Long): QueryReplacementClassDetailsResponse
-}
