--- PlanIt-application/src/main/kotlin/com/nodoubt/planitbackend/domain/replacementClass/api/dto/response/QueryReplacementClassListResponse.kt
@@ -1,7 +1,0 @@
-package com.nodoubt.planitbackend.domain.replacementClass.api.dto.response
-
-import com.nodoubt.planitbackend.domain.replacementClass.spi.vo.ReplacementClassListVO
-
-data class QueryReplacementClassListResponse(
-    val replacementClassList: List<ReplacementClassListVO>
-)
