--- PlanIt-application/src/main/kotlin/com/nodoubt/planitbackend/domain/replacementClass/spi/CommandReplacementPort.kt
@@ -1,7 +1,0 @@
-package com.nodoubt.planitbackend.domain.replacementClass.spi
-
-import com.nodoubt.planitbackend.domain.replacementClass.domain.ReplacementClass
-
-interface CommandReplacementPort {
-    fun saveReplacementClassList(replacementClass: List<ReplacementClass>)
-}
