--- PlanIt-application/src/main/kotlin/com/nodoubt/planitbackend/domain/replacementClass/spi/ReplacementClassPort.kt
@@ -1,3 +1,0 @@
-package com.nodoubt.planitbackend.domain.replacementClass.spi
-
-interface ReplacementClassPort : CommandReplacementPort, QueryReplacementClassPort
