+++ PlanIt-infrastructure/src/main/kotlin/com/nodoubt/planitbackend/domain/changeMaster/presentation/ChangeMasterWebAdapter.kt
@@ -1,23 +1,25 @@
 package com.nodoubt.planitbackend.domain.changeMaster.presentation
 
 import com.nodoubt.planitbackend.domain.changeMaster.api.ChangeTimetablePort
+import com.nodoubt.planitbackend.domain.changeMaster.api.QueryReplacementClassDetailsPort
+import com.nodoubt.planitbackend.domain.changeMaster.api.QueryReplacementClassListPort
 import com.nodoubt.planitbackend.domain.changeMaster.api.QueryRequestedTimetableListPort
+import com.nodoubt.planitbackend.domain.changeMaster.api.dto.response.QueryReplacementClassDetailsResponse
+import com.nodoubt.planitbackend.domain.changeMaster.api.dto.response.QueryReplacementClassListResponse
 import com.nodoubt.planitbackend.domain.changeMaster.api.dto.response.QueryRequestedTimetableListResponse
+import com.nodoubt.planitbackend.domain.changeMaster.domain.Status
 import com.nodoubt.planitbackend.domain.changeMaster.presentation.dto.request.WebChangeTimetableRequest
 import org.springframework.http.HttpStatus
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.ResponseStatus
-import org.springframework.web.bind.annotation.RestController
+import org.springframework.web.bind.annotation.*
 import javax.validation.Valid
 
 @RestController
 @RequestMapping("/timetables")
 class ChangeMasterWebAdapter(
     private val changeTimetablePort: ChangeTimetablePort,
-    private val queryRequestedTimetableListPort: QueryRequestedTimetableListPort
+    private val queryRequestedTimetableListPort: QueryRequestedTimetableListPort,
+    private val queryReplacementClassListPort: QueryReplacementClassListPort,
+    private val queryReplacementClassDetailsPort: QueryReplacementClassDetailsPort
 ) {
 
     @ResponseStatus(HttpStatus.CREATED)
@@ -29,4 +31,18 @@ class ChangeMasterWebAdapter(
     @GetMapping("/requested")
     fun queryRequestedTimetableList(): QueryRequestedTimetableListResponse =
         queryRequestedTimetableListPort.execute()
+
+    @GetMapping("/replacement-class")
+    fun queryReplacementClassList(
+        @RequestParam("status", required = false) status: Status?
+    ): QueryReplacementClassListResponse {
+        return queryReplacementClassListPort.execute(status)
+    }
+
+    @GetMapping("/replacement-class/{replacement-class-id}")
+    fun queryReplacementClassDetails(
+        @PathVariable("replacement-class-id") replacementClassId: Long
+    ): QueryReplacementClassDetailsResponse {
+        return queryReplacementClassDetailsPort.execute(replacementClassId)
+    }
 }
