+++ PlanIt-infrastructure/src/main/kotlin/com/nodoubt/planitbackend/domain/replacementClass/persistence/vo/QueryReplacementClassListVO.kt
@@ -1,7 +1,7 @@
 package com.nodoubt.planitbackend.domain.replacementClass.persistence.vo
 
 import com.nodoubt.planitbackend.domain.changeMaster.domain.Status
-import com.nodoubt.planitbackend.domain.replacementClass.spi.vo.ReplacementClassListVO
+import com.nodoubt.planitbackend.domain.changeMaster.spi.replacementClassSpi.vo.ReplacementClassListVO
 import com.querydsl.core.annotations.QueryProjection
 import java.time.LocalDate
